+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## v0.0.14 - in progress
-- Human-readable report describing which fields show up in which schemas.
+- Add Excel sheet describing which fields show up in which schemas.
 
 ## v0.0.13 - 2022-01-07
 - Make more fields explicitly numeric